ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
numpy==1.26.2
scikit-learn==1.3.2
gunicorn==21.2.0

